+++ omniload/source/excel/api.py
@@ -1,0 +1,11 @@
+class ExcelSource:
+    def handles_incrementality(self) -> bool:
+        return True
+
+    def dlt_source(self, uri: str, table: str, **kwargs):
+
+        path = uri.replace("excel://", "")
+
+        from omniload.source.excel.adapter import excel_source
+
+        return excel_source(path, table)