+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## in progress
 
+- Connectors: Added source adapter for the XLSX file format
+
 ## 2026/07/02 v0.5.0
 
 - Connectors: Updated to Asana client v5